Prøver meg på en liten cheeky

+++ trianglesurface.h
@@ -19,7 +19,7 @@ public:
     void construct_plane();
     std::vector<gsml::Vertex>& get_vertices() { return mVertices; } // 191120
     float xmin=0, xmax=1.0f, ymin=0.0f, ymax=1.0f, h=0.25f;
-    double xMinimum = 0, xMaximum = 1, yMinimum = 0, yMaximum=1, zMinimum = 0, zMaximum=1;
+    float xMinimum = 0, xMaximum = 1, yMinimum = 0, yMaximum=1, zMinimum = 0, zMaximum=1;
 };
 
 #endif // TRIANGLESURFACE_H
